Refactor and cleanup editor
